requirements.txt
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
streamlit>=1.40.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
